Variables can now be connected to nodes
This is done using a multiselect. Multiple variables can be selected at once. Variables are then added to an accordion in the modal popup. Currently only the json data can be seen, but later on this can be used for editing the representation of the variable. 
The number of times a variable is used is also counted.