Some layout fixes, added colorpicker for nodes
Had some trouble with the sidebar (particulary in mobile view). Replaced malihu-scrollbar plugin with yasimplescrollbar (may need to be changed) to fix this. Made title in navbar change based on if it fit or not, though the title is now hardcoded. Fixed tiny bugs. Added colorpicker (using palette) to options-modal, mostly so that, when this is added, the children steps can be distinguishable.