Nav-tabs for in modal, weird visual glitches fixed
ID of step is now shown inside step, name could be done as well but id is cleaner. Right now id is always -1, will be id in database to keep steps apart from eachother. The content of a navtab has a fixed height, overflow will scroll. For some reason the scrollbar doesn't show up, though it is scrollable. Not sure why.